Allow rendering of axis titles inside or outside of plots dimensions.
Adds a function setIncludeAxisLabelsInPlotDimensions(bool x, bool y)
that allows the user to specify whether the axis titles (which are
rendered outside of the plot itself) should be counted as part of the
plot’s dimensions (x and width, y and height) or rendered outside of
them.

Fixes https://github.com/nickgillian/ofxGrt/issues/18.